% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_colnames.R
\name{compress_header}
\alias{compress_header}
\title{Compress Column Names}
\usage{
compress_header(header_rows)
}
\arguments{
\item{header_rows}{numeric. The number of rows that the column names span.}
}
\value{
a character vector of collapsed column names
}
\description{
Compress Column Names
}
