% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{percent_missing}
\alias{percent_missing}
\title{Determine Row Percentage NA Values}
\usage{
percent_missing(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric vector containing the percentage of NA values in x
}
\description{
Determine Row Percentage NA Values
}
