% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data From a Report}
\usage{
extract_data(.df, checkcol = NA, promote_colnames = TRUE)
}
\arguments{
\item{.df}{a data frame object.}

\item{checkcol}{numeric. (optional). The column number to check for NA values.
If not specified, extract_data will attempt to guess the appropriate column.}

\item{promote_colnames}{logical. (optional). Should the first row be promoted
to column names?}
}
\value{
a data frame object.
}
\description{
Extract Data From a Report
}
\examples{
df <- tribble(~V0,     ~v1,           ~v2,      ~v3,
              "Report on values", NA,  NA,       NA,
               NA,      "X",           "Y",      "Z",
               NA, "1234.5",      "234.12", "3452.1")
extract_data(df)
}
