% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_colnames.R
\name{promote_colnames}
\alias{promote_colnames}
\title{Promote Rows to Column Names}
\usage{
promote_colnames(.df, header_start = 1, header_nrow = 1)
}
\arguments{
\item{.df}{a data frame}

\item{header_start}{numeric. The first row of the column headers to be promoted.}

\item{header_nrow}{numeric. The number of rows the column headers span. Defaults to 1.}
}
\value{
a data frame with the promoted row as column names
}
\description{
Promote Rows to Column Names
}
\examples{
df <- data.frame(v1 = c("X", "123.45"), v2 = c("Y", "345.67"))
promote_colnames(df)
df2 <- data.frame(v1 = c("X", "Value", "123.45"), v2 = c("Y", "Value", "345.67"))
promote_colnames(df2, header_nrow = 2)
}
