% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_colnames.R
\name{promote_colnames}
\alias{promote_colnames}
\title{Promote Row to Column Names}
\usage{
promote_colnames(.df, rownum = 1)
}
\arguments{
\item{.df}{a data frame}

\item{rownum}{numeric. The row in the data frame to be promoted to column names}
}
\value{
a data frame with the promoted row as column names
}
\description{
Promote Row to Column Names
}
\examples{
df <- data.frame(v1 = c("X", "123.45"), v2 = c("Y", "345.67"))
promote_colnames(df)
}
