% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_metadata.R
\name{strip_metadata}
\alias{strip_metadata}
\title{Strip Report MetaData From Excel File}
\usage{
strip_metadata(from, promote = NA, header_nrow = 1, ...)
}
\arguments{
\item{from}{the file path to the excel report.}

\item{promote}{logical. Should first non-NA row be promoted to column names?
Default tries to make a reasonable determination.}

\item{header_nrow}{numeric. How many rows does the table header cover? Defaults to 1.}

\item{...}{extra parameters to pass to the read_csv and read_excel functions}
}
\value{
a data frame containing the tabular data from the report
}
\description{
Strip Report MetaData From Excel File
}
\examples{
\dontrun{
strip_metadata("./man/example/example-report")
# specify a particular sheet
strip_metadata("my-excel-file.xlsx", sheet = "report summary")
}
}
