% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_metadata.R
\name{strip_metadata}
\alias{strip_metadata}
\title{Strip Report MetaData From Excel File}
\usage{
strip_metadata(from, ...)
}
\arguments{
\item{from}{the file path to the excel report.}

\item{...}{extra parameters to pass to the extract_data function}
}
\value{
a data frame containing the tabular data from the report
}
\description{
Strip Report MetaData From Excel File
}
\examples{
\dontrun{
strip_metadata("./man/example/example-report")
}
}
